PR fortran/95091 - Buffer overflows with submodules and long symbols
Add cast to fix bootstrap error with -Werror=sign-compare.

gcc/fortran/
	PR fortran/95091
	* class.c (gfc_hash_value): Add cast.

(cherry picked from commit 5aaccde3db39fac7e7f6677ceccc1eadd9c6a424)

+++ gcc/fortran/class.c
@@ -532,7 +532,7 @@ gfc_hash_value (gfc_symbol *sym)
 
   get_unique_type_string (&c[0], sym);
   len = strnlen (c, sizeof (c));
-  gcc_assert (len < sizeof (c));
+  gcc_assert ((size_t) len < sizeof (c));
 
   for (i = 0; i < len; i++)
     hash = (hash << 6) + (hash << 16) - hash + c[i];
